feat:! small fixes
